+++ WagoUI_Creator/modules/generic.lua
@@ -4,6 +4,7 @@ local LAP = LibStub:GetLibrary("LibAddonProfiles")
 
 local defaultSortOrder = {
   "Additional Addons",
+  "UI Scale",
   "Blizzard Edit Mode",
   "Blizzard Cooldown Manager",
   "ElvUI",